fix: Prevent zero-ing out public keys on updates
Due to the fact that the KeyGeneratorFunction did not handle UPDATE events, the downstream CloudFrontSigningPublicKey resource would receive an empty EncodedKey an attempt to update the resource to an invalid state.  As a fix, we have update the keypair generation logic to return the stored public key on updates, ensuring that downstream resources would receive the same output on UPDATE events as they do on CREATE events. This required us to store the public key in SSM on create.  Along the way, we've simplified the function payloads within the custom resource to promote a more unified format between create and update events. Additionally, we've removed unnecessary permissions.